Made sure the binary expire value was correct
Before an expire value of eg. 86400 didn't work while an expire value of eg. 7200 did work. 86400 resulted in an "invalid argument" responsefrom Memcached. Tracked this down to the expire value being corrupted in an unnecessary conversion between its binary buffer value and an UTF-8 string values. Removed the unnecessary conversion and kept the binary value in the buffer all the way and both 86400 and 7200 now works.