Fix bug where we write to socket before auth finishes.
If two different requests handlers are run at the same time, the first
one will start the authentication of the socket, while the second will
try to write it's request to the socket. The second write hits the
memcache server first and causes the connection to be reset.

This fixes issue #65.

We could presumably also fix this by just sending the SASL auth first
thing and not bothering with `listSasl`, but that's more of a hack
anyway, so I prefer this approach.