New timeout handling that uses a single active timer
We move to a new (yet again!) timeout handling system where we only use
one active timer at a time, and track future timeouts / deadlines in our
own queue to re-arm the timer.

This fixes issue #52.

We could use `setMaxListeners` to instead increase the timeout handlers
allowed and keep with the existing design, but this approach is O(1)
scalable.