Switch to new argument scheme with options dictionary
We now declare methods generally as:

        function(key, value, options, callback)

Rather than adding each argument (generally unused) as trailing
arguments after callback. Callback is _always_ the last argument now.

To handle backwards compatability, we check the type of `options` and if
it is a function, we treat arguments as corresponding to the old
interface but issue a depreceated warning.

This fixes issue #53 and #50, making CAS also easier to support going
forward.