Fixes set(...) not backwards compatible on null callback.
Clients still using the old argument ordering with a null or undefined callback are broken since the API determines old vs new usage using `typeof options` for which null is a valid case. This fix also detects the old API usage in the case that callback is the expiration interval.