Only call socket error callback on `close` event
The `close` event is always called after an `error` event (https://nodejs.org/api/net.html#event-error_1), and `disconnect` is not an event emitted from a `net.Socket`. So we should only wait for a `close` event (which captures both `error` cases and non-`error` cases such as the server cleaning closing the connection with a FIN). This will avoid calling error callbacks twice (I believe this isn't technically a problem at the moment, but only calling `Server#error` once for a close should be more maintainable).